requirements.txt
Django==1.11.4
django_mptt==0.8.7
Pillow==4.2.1
Jinja2==2.9.6